refactor: extract curved free-disk shape-gradient projections
- Move specialized curved/free-disk shape-gradient projection helpers from runtime/minimizer.py to runtime/projections/curved_disk.py.
- Update the minimizer loop to call the module-level projection function.
- Preserve exact call order: projection still happens after project_constraints_array(grad_arr).
- Preserve in-place grad_arr mutation and existing global_params gates.
